scripts/requirements.txt
pandas>=2.2.0
python-dotenv>=1.0.0
openpyxl>=3.1.5
